refactor: pair ~56 ABI-identical signatures with the binary (close coverage holes)
Wave 2: these functions were byte-identical on the ARM32 cross-build but
their MANGLED names differed (const&/qualifier, or int<->long width), so
asm-verify -- which pairs on exact mangled name -- never diffed their
bodies. They were silent coverage holes, not "verified MATCH". Aligning
the port signatures to the binary makes asm-verify pair + diff them, so
future body drift surfaces (the class of gap this session repeatedly
found hiding real bugs).

All mechanical, no behavior change (ARM32 passes const& and by-value
aggregates the same; long==int==4B on both ARM32 and Windows x64):
- const& -> by-value where the binary passes by value (Colour, Vec3,
  Delegate0, ...).
- const method-qualifier added/removed to match the binary.
- int/unsigned int -> long/unsigned long width where the binary uses long.
- A handful of const& -> non-const& (Math::ClosestPointOnLine/LineIntersect,
  FancyBakedStringBuild) with local-copy/const_cast call-site fixups.
- Game::Init/MortarGame::Init -> (int, const char**) (coupled override).

Skipped (deferred): 6 SmartPtr/AsciiString by-value cases (host copy
cost), the 7 ES2 Draw-refactor + 5 templated/std findings (intentional /
noise), and the 2 D0/D2 virtual-destructor-variant mismatches
(Game::~Game, NetworkManager::~NetworkManager -- vtable-layout question,
need re-analyst). Host + full-game build green.

+++ src/engine/input/Touch.h
@@ -110,18 +110,18 @@ public:
     // Push TEvnt to ring; on overflow: Update(0.0f) then retry.
     // SDL entry point: InputTranslatorSDL calls this for each touch event.
     // isActive: true=press OR move, false=release (matches binary param 'b').
-    void __UpdateInternal(uint32_t extId, bool isActive, float x, float y, float t);
+    void __UpdateInternal(unsigned long extId, bool isActive, float x, float y, float t);
 
     // Binary @ 0x00195314 -- ___UpdateInternal.
     // Apply event to states2: match by extId or claim free slot (rotating cursor).
     // nextTouchId++ skipping 0 on wrap.
     // isActive: true=press OR move, false=release (matches binary param 'b').
     // Binary @ 0x00195314 -- free slot is extId==0, NOT phase>=1.
-    void ___UpdateInternal(uint32_t extId, bool isActive, float x, float y);
+    void ___UpdateInternal(unsigned long extId, bool isActive, float x, float y);
 
     // Binary @ 0x00195424 -- FindTouch(uint touchId).
     // Linear scan states1; return slot index or -1.
-    int FindTouch(uint32_t touchId);
+    int FindTouch(unsigned long touchId);
 
     // Binary @ 0x001954fc -- GetAnyTouch().
     // First slot with phase < 1; returns touchId or 0.
@@ -134,11 +134,11 @@ public:
     // ASM-spec v1.6.1 Touch::GetTouchPos @0x002429d4: (uint, float&, float&).
     // Writes currX/Y of matching slot. Returns 1 if active (phase < 1), 0 if not.
     // Binary leaves *x/*y UNTOUCHED on miss.
-    int GetTouchPos(uint32_t touchId, float& x, float& y);
+    int GetTouchPos(unsigned long touchId, float& x, float& y);
 
     // ASM-spec v1.6.1 Touch::GetTouchDelta @0x00242a20: (uint, float&, float&).
     // Writes currX-prevX/dy if phase >= 0, else 0.0f. Returns 1 if active.
-    int GetTouchDelta(uint32_t touchId, float& dx, float& dy);
+    int GetTouchDelta(unsigned long touchId, float& dx, float& dy);
 
     // Binary @ 0x00242a98 (v1.6.1) -- GetTouchInReigion (note binary typo).
     // Find first active touch inside (x, y, x+w, y+h). Returns touchId or 0.
