+++ src/entities/Entity.h
@@ -116,7 +116,7 @@ public:
     // Called from GameInit step 15 with 0x20000 (128 KB) to allocate the
     // process-global LinkedHeap Entity arena before ActorManager::Initialise.
     // DIFFERS: original = LinkedHeap arena 0x20000, port uses std new (no fixed cap).
-    static void HeapCreate(unsigned int bytes);
+    static void HeapCreate(unsigned long bytes);
 
     // Counterpart to HeapCreate; called from GameExit.
     // Binary: Entity::HeapDestroy @ 0x0019d6d0.
